+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/db/migration/mariadb/V13__Remove_Task2_Batch4_Support.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.dataflow.server.db.migration.sqlserver;
+package org.springframework.cloud.dataflow.server.db.migration.mariadb;
 
 import java.util.Arrays;
 import java.util.List;
@@ -26,7 +26,7 @@ import org.springframework.cloud.dataflow.server.db.migration.AbstractRemoveBatc
  *
  * @author Glenn Renfro
  */
-public class V10__Remove_Task2_Batch4_Support extends AbstractRemoveBatch4Task2Tables {
+public class V13__Remove_Task2_Batch4_Support extends AbstractRemoveBatch4Task2Tables {
 
 	/*
 	 * Scripts to remove views used for Task V2/Batch V4 Task V3/Batch V5 queries.
@@ -53,111 +53,111 @@ public class V10__Remove_Task2_Batch4_Support extends AbstractRemoveBatch4Task2T
 	 * Scripts to rename Task V2 tables removing BOOT_ prefix.
 	 */
 	private final static String RENAME_TASK_EXECUTION_V2_TABLE =
-		"exec sp_rename 'TASK_EXECUTION', 'V2_TASK_EXECUTION'";
+		"ALTER TABLE TASK_EXECUTION RENAME TO V2_TASK_EXECUTION";
 
 	private final static String RENAME_TASK_EXECUTION_PARAMS_V2_TABLE =
-		"exec sp_rename 'TASK_EXECUTION_PARAMS', 'V2_TASK_EXECUTION_PARAMS'";
+		"ALTER TABLE TASK_EXECUTION_PARAMS RENAME TO V2_TASK_EXECUTION_PARAMS";
 
 	private final static String RENAME_TASK_TASK_BATCH_V2_TABLE =
-		"exec sp_rename 'TASK_TASK_BATCH', 'V2_TASK_TASK_BATCH'";
+		"ALTER TABLE TASK_TASK_BATCH RENAME TO V2_TASK_TASK_BATCH";
 
 	private final static String RENAME_TASK_LOCK_V2_TABLE =
-		"exec sp_rename 'TASK_LOCK', 'V2_TASK_LOCK'";
+		"ALTER TABLE TASK_LOCK RENAME TO V2_TASK_LOCK";
 
 	private final static String RENAME_TASK_V2_SEQ =
-		"exec sp_rename 'TASK_SEQ', 'V2_TASK_SEQ'";
+		"ALTER TABLE TASK_SEQ RENAME TO V2_TASK_SEQ";
 
 	private final static String RENAME_TASK_EXECUTION_METADATA_V2_TABLE =
-		"exec sp_rename 'TASK_EXECUTION_METADATA', 'V2_TASK_EXECUTION_METADATA'";
+		"ALTER TABLE TASK_EXECUTION_METADATA RENAME TO V2_TASK_EXECUTION_METADATA";
 
 	private final static String RENAME_TASK_EXECUTION_METADATA_V2_SEQ =
-		"exec sp_rename 'TASK_EXECUTION_METADATA_SEQ', 'V2_TASK_EXECUTION_METADATA_SEQ'";
+		"ALTER TABLE TASK_EXECUTION_METADATA_SEQ RENAME TO V2_TASK_EXECUTION_METADATA_SEQ";
 
 	/*
 	 * Scripts to rename Batch V5 tables removing BOOT_ prefix.
 	 */
 
 	private final static String RENAME_BATCH_JOB_INSTANCE_V4_TABLE =
-		"exec sp_rename 'BATCH_JOB_INSTANCE', 'V2_BATCH_JOB_INSTANCE'";
+		"ALTER TABLE BATCH_JOB_INSTANCE RENAME TO V2_BATCH_JOB_INSTANCE";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_V4_TABLE =
-		"exec sp_rename 'BATCH_JOB_EXECUTION', 'V2_BATCH_JOB_EXECUTION'";
+		"ALTER TABLE BATCH_JOB_EXECUTION RENAME TO V2_BATCH_JOB_EXECUTION";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_PARAMS_V4_TABLE =
-		"exec sp_rename 'BATCH_JOB_EXECUTION_PARAMS', 'V2_BATCH_JOB_EXECUTION_PARAMS'";
+		"ALTER TABLE BATCH_JOB_EXECUTION_PARAMS RENAME TO V2_BATCH_JOB_EXECUTION_PARAMS";
 
 	private final static String RENAME_BATCH_STEP_EXECUTION_V4_TABLE =
-		"exec sp_rename 'BATCH_STEP_EXECUTION', 'V2_BATCH_STEP_EXECUTION'";
+		"ALTER TABLE BATCH_STEP_EXECUTION RENAME TO V2_BATCH_STEP_EXECUTION";
 
 	private final static String RENAME_BATCH_STEP_EXECUTION_CONTEXT_V4_TABLE =
-		"exec sp_rename 'BATCH_STEP_EXECUTION_CONTEXT', 'V2_BATCH_STEP_EXECUTION_CONTEXT'";
+		"ALTER TABLE BATCH_STEP_EXECUTION_CONTEXT RENAME TO V2_BATCH_STEP_EXECUTION_CONTEXT";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_CONTEXT_V4_TABLE =
-		"exec sp_rename 'BATCH_JOB_EXECUTION_CONTEXT', 'V2_BATCH_JOB_EXECUTION_CONTEXT'";
+		"ALTER TABLE BATCH_JOB_EXECUTION_CONTEXT RENAME TO V2_BATCH_JOB_EXECUTION_CONTEXT";
 
 	private final static String RENAME_BATCH_STEP_EXECUTION_V4_SEQ =
-		"exec sp_rename 'BATCH_STEP_EXECUTION_SEQ', 'V2_BATCH_STEP_EXECUTION_SEQ'";
+		"ALTER TABLE BATCH_STEP_EXECUTION_SEQ RENAME TO V2_BATCH_STEP_EXECUTION_SEQ";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_V4_SEQ =
-		"exec sp_rename 'BATCH_JOB_EXECUTION_SEQ', 'V2_BATCH_JOB_EXECUTION_SEQ'";
+		"ALTER TABLE BATCH_JOB_EXECUTION_SEQ RENAME TO V2_BATCH_JOB_EXECUTION_SEQ";
 
 	private final static String RENAME_BATCH_JOB_V4_SEQ =
-		"exec sp_rename 'BATCH_JOB_SEQ', 'V2_BATCH_JOB_SEQ'";
+		"ALTER TABLE BATCH_JOB_SEQ RENAME TO V2_BATCH_JOB_SEQ";
 
 	/*
 	 * Scripts to rename Task V3 tables removing BOOT_ prefix.
 	 */
 	private final static String RENAME_TASK_EXECUTION_V3_TABLE =
-			"exec sp_rename 'BOOT3_TASK_EXECUTION', 'TASK_EXECUTION'";
+			"ALTER TABLE BOOT3_TASK_EXECUTION RENAME TO TASK_EXECUTION";
 
 	private final static String RENAME_TASK_EXECUTION_PARAMS_V3_TABLE =
-			"exec sp_rename 'BOOT3_TASK_EXECUTION_PARAMS', 'TASK_EXECUTION_PARAMS'";
+			"ALTER TABLE BOOT3_TASK_EXECUTION_PARAMS RENAME TO TASK_EXECUTION_PARAMS";
 
 	private final static String RENAME_TASK_TASK_BATCH_V3_TABLE =
-			"exec sp_rename 'BOOT3_TASK_TASK_BATCH', 'TASK_TASK_BATCH'";
+			"ALTER TABLE BOOT3_TASK_TASK_BATCH RENAME TO TASK_TASK_BATCH";
 
 	private final static String RENAME_TASK_LOCK_V3_TABLE =
-			"exec sp_rename 'BOOT3_TASK_LOCK', 'TASK_LOCK'";
+			"ALTER TABLE BOOT3_TASK_LOCK RENAME TO TASK_LOCK";
 
 	private final static String RENAME_TASK_V3_SEQ =
-			"exec sp_rename 'BOOT3_TASK_SEQ', 'TASK_SEQ'";
+			"ALTER TABLE BOOT3_TASK_SEQ RENAME TO TASK_SEQ";
 
 	private final static String RENAME_TASK_EXECUTION_METADATA_V3_TABLE =
-			"exec sp_rename 'BOOT3_TASK_EXECUTION_METADATA', 'TASK_EXECUTION_METADATA'";
+			"ALTER TABLE BOOT3_TASK_EXECUTION_METADATA RENAME TO TASK_EXECUTION_METADATA";
 
 	private final static String RENAME_TASK_EXECUTION_METADATA_V3_SEQ =
-			"exec sp_rename 'BOOT3_TASK_EXECUTION_METADATA_SEQ', 'TASK_EXECUTION_METADATA_SEQ'";
+			"ALTER TABLE BOOT3_TASK_EXECUTION_METADATA_SEQ RENAME TO TASK_EXECUTION_METADATA_SEQ";
 
 	/*
 	 * Scripts to rename Batch V5 tables removing BOOT_ prefix.
 	 */
 
 	private final static String RENAME_BATCH_JOB_INSTANCE_V5_TABLE =
-			"exec sp_rename 'BOOT3_BATCH_JOB_INSTANCE', 'BATCH_JOB_INSTANCE'";
+			"ALTER TABLE BOOT3_BATCH_JOB_INSTANCE RENAME TO BATCH_JOB_INSTANCE";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_V5_TABLE =
-			"exec sp_rename 'BOOT3_BATCH_JOB_EXECUTION', 'BATCH_JOB_EXECUTION'";
+			"ALTER TABLE BOOT3_BATCH_JOB_EXECUTION RENAME TO BATCH_JOB_EXECUTION";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_PARAMS_V5_TABLE =
-			"exec sp_rename 'BOOT3_BATCH_JOB_EXECUTION_PARAMS', 'BATCH_JOB_EXECUTION_PARAMS'";
+			"ALTER TABLE BOOT3_BATCH_JOB_EXECUTION_PARAMS RENAME TO BATCH_JOB_EXECUTION_PARAMS";
 
 	private final static String RENAME_BATCH_STEP_EXECUTION_V5_TABLE =
-			"exec sp_rename 'BOOT3_BATCH_STEP_EXECUTION', 'BATCH_STEP_EXECUTION'";
+			"ALTER TABLE BOOT3_BATCH_STEP_EXECUTION RENAME TO BATCH_STEP_EXECUTION";
 
 	private final static String RENAME_BATCH_STEP_EXECUTION_CONTEXT_V5_TABLE =
-			"exec sp_rename 'BOOT3_BATCH_STEP_EXECUTION_CONTEXT', 'BATCH_STEP_EXECUTION_CONTEXT'";
+			"ALTER TABLE BOOT3_BATCH_STEP_EXECUTION_CONTEXT RENAME TO BATCH_STEP_EXECUTION_CONTEXT";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_CONTEXT_V5_TABLE =
-			"exec sp_rename 'BOOT3_BATCH_JOB_EXECUTION_CONTEXT', 'BATCH_JOB_EXECUTION_CONTEXT'";
+			"ALTER TABLE BOOT3_BATCH_JOB_EXECUTION_CONTEXT RENAME TO BATCH_JOB_EXECUTION_CONTEXT";
 
 	private final static String RENAME_BATCH_STEP_EXECUTION_V5_SEQ =
-			"exec sp_rename 'BOOT3_BATCH_STEP_EXECUTION_SEQ', 'BATCH_STEP_EXECUTION_SEQ'";
+			"ALTER TABLE BOOT3_BATCH_STEP_EXECUTION_SEQ RENAME TO BATCH_STEP_EXECUTION_SEQ";
 
 	private final static String RENAME_BATCH_JOB_EXECUTION_V5_SEQ =
-			"exec sp_rename 'BOOT3_BATCH_JOB_EXECUTION_SEQ', 'BATCH_JOB_EXECUTION_SEQ'";
+			"ALTER TABLE BOOT3_BATCH_JOB_EXECUTION_SEQ RENAME TO BATCH_JOB_EXECUTION_SEQ";
 
 	private final static String RENAME_BATCH_JOB_V5_SEQ =
-			"exec sp_rename 'BOOT3_BATCH_JOB_SEQ', 'BATCH_JOB_SEQ'";
+			"ALTER TABLE BOOT3_BATCH_JOB_SEQ RENAME TO BATCH_JOB_SEQ";
 
 	@Override
 	public List<SqlCommand> dropBoot3Boot2Views() {
