+++ spring-cloud-dataflow-server-core/src/main/resources/application-init-db2.yml
@@ -16,4 +16,5 @@ spring:
         - classpath*:/schemas/db2/V7-dataflow.sql
         - classpath*:/schemas/db2/V8-dataflow.sql
         - classpath*:/schemas/db2/V9-dataflow.sql
+        - classpath*:/schemas/db2/V10-dataflow.sql
