+++ spring-cloud-dataflow-server-core/src/main/resources/application-init-mysql.yml
@@ -18,3 +18,4 @@ spring:
         - classpath*:/schemas/mysql/V8-dataflow.sql
         - classpath*:/schemas/mysql/V9-dataflow.sql
         - classpath*:/schemas/mysql/V10-dataflow.sql
+        - classpath*:/schemas/mysql/V11-dataflow.sql
