+++ spring-cloud-dataflow-server-core/src/main/resources/application-init-postgresql.yml
@@ -17,3 +17,4 @@ spring:
         - classpath*:/schemas/postgresql/V8-dataflow.sql
         - classpath*:/schemas/postgresql/V9-dataflow.sql
         - classpath*:/schemas/postgresql/V10-dataflow.sql
+        - classpath*:/schemas/postgresql/V11-dataflow.sql
