+++ spring-cloud-dataflow-server-core/src/main/resources/application-init-sqlserver.yml
@@ -17,3 +17,4 @@ spring:
         - classpath*:/schemas/sqlserver/V7-dataflow.sql
         - classpath*:/schemas/sqlserver/V8-dataflow.sql
         - classpath*:/schemas/sqlserver/V9-dataflow.sql
+        - classpath*:/schemas/sqlserver/V10-dataflow.sql
