+++ spring-cloud-dataflow-server-core/src/main/resources/schemas/postgresql/V11-dataflow.sql
@@ -1,0 +1,72 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *          http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+/**
+ * Remove aggregate views
+ */
+
+DROP VIEW AGGREGATE_TASK_EXECUTION;
+DROP VIEW AGGREGATE_TASK_EXECUTION_PARAMS;
+DROP VIEW AGGREGATE_JOB_EXECUTION;
+DROP VIEW AGGREGATE_JOB_INSTANCE;
+DROP VIEW AGGREGATE_TASK_BATCH;
+DROP VIEW AGGREGATE_STEP_EXECUTION;
+
+/*
+ * Prefix Task V2 and Batch V4 tables with V2_ prefix.
+ * Allow user to determine what they should do with these tables.
+ */
+ALTER TABLE TASK_EXECUTION RENAME TO V2_TASK_EXECUTION;
+ALTER TABLE TASK_EXECUTION_PARAMS RENAME TO V2_TASK_EXECUTION_PARAMS;
+ALTER TABLE TASK_TASK_BATCH RENAME TO V2_TASK_TASK_BATCH;
+ALTER TABLE TASK_LOCK RENAME TO V2_TASK_LOCK;
+ALTER TABLE TASK_SEQ RENAME TO V2_TASK_SEQ;
+ALTER TABLE TASK_EXECUTION_METADATA RENAME TO V2_TASK_EXECUTION_METADATA;
+ALTER TABLE TASK_EXECUTION_METADATA_SEQ RENAME TO V2_TASK_EXECUTION_METADATA_SEQ;
+ALTER TABLE BATCH_JOB_INSTANCE RENAME TO V2_BATCH_JOB_INSTANCE;
+ALTER TABLE BATCH_JOB_EXECUTION RENAME TO V2_BATCH_JOB_EXECUTION;
+ALTER TABLE BATCH_JOB_EXECUTION_PARAMS RENAME TO V2_BATCH_JOB_EXECUTION_PARAMS;
+ALTER TABLE BATCH_STEP_EXECUTION RENAME TO V2_BATCH_STEP_EXECUTION;
+ALTER TABLE BATCH_STEP_EXECUTION_CONTEXT RENAME TO V2_BATCH_STEP_EXECUTION_CONTEXT;
+ALTER TABLE BATCH_JOB_EXECUTION_CONTEXT RENAME TO V2_BATCH_JOB_EXECUTION_CONTEXT;
+ALTER TABLE BATCH_STEP_EXECUTION_SEQ RENAME TO V2_BATCH_STEP_EXECUTION_SEQ;
+ALTER TABLE BATCH_JOB_EXECUTION_SEQ RENAME TO V2_BATCH_JOB_EXECUTION_SEQ;
+ALTER TABLE BATCH_JOB_SEQ RENAME TO V2_BATCH_JOB_SEQ;
+
+
+/*
+ * Remove BOOT3_ prefix Task V3 and Batch V5 tables .
+ */
+
+ALTER TABLE BOOT3_TASK_EXECUTION RENAME TO TASK_EXECUTION;
+ALTER TABLE BOOT3_TASK_EXECUTION_PARAMS RENAME TO TASK_EXECUTION_PARAMS;
+ALTER TABLE BOOT3_TASK_TASK_BATCH RENAME TO TASK_TASK_BATCH;
+ALTER TABLE BOOT3_TASK_LOCK RENAME TO TASK_LOCK;
+ALTER TABLE BOOT3_TASK_SEQ RENAME TO TASK_SEQ;
+ALTER TABLE BOOT3_TASK_EXECUTION_METADATA RENAME TO TASK_EXECUTION_METADATA;
+ALTER TABLE BOOT3_TASK_EXECUTION_METADATA_SEQ RENAME TO TASK_EXECUTION_METADATA_SEQ;
+ALTER TABLE BOOT3_BATCH_JOB_INSTANCE RENAME TO BATCH_JOB_INSTANCE;
+ALTER TABLE BOOT3_BATCH_JOB_EXECUTION RENAME TO BATCH_JOB_EXECUTION;
+ALTER TABLE BOOT3_BATCH_JOB_EXECUTION_PARAMS RENAME TO BATCH_JOB_EXECUTION_PARAMS;
+ALTER TABLE BOOT3_BATCH_STEP_EXECUTION RENAME TO BATCH_STEP_EXECUTION;
+ALTER TABLE BOOT3_BATCH_STEP_EXECUTION_CONTEXT RENAME TO BATCH_STEP_EXECUTION_CONTEXT;
+ALTER TABLE BOOT3_BATCH_JOB_EXECUTION_CONTEXT RENAME TO BATCH_JOB_EXECUTION_CONTEXT;
+ALTER TABLE BOOT3_BATCH_STEP_EXECUTION_SEQ RENAME TO BATCH_STEP_EXECUTION_SEQ;
+ALTER TABLE BOOT3_BATCH_JOB_EXECUTION_SEQ RENAME TO BATCH_JOB_EXECUTION_SEQ;
+ALTER TABLE BOOT3_BATCH_JOB_SEQ RENAME TO BATCH_JOB_SEQ;
+
+
+
